requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=1.0.0
anthropic>=0.39.0
aiohttp>=3.9.0
